refactor: share API key metadata parsing
